Split PlotItems into an items package + fix visible = False rendering bug
